Create separate directory for time range preset components and unify naming.
