+++ graylog2-web-interface/src/views/components/searchbar/TimeRangeInput.tsx
@@ -24,10 +24,10 @@ import type { TimeRange, NoTimeRangeOverride } from 'views/logic/queries/Query';
 import { SEARCH_BAR_GAP } from 'views/components/searchbar/SearchBarLayout';
 import useSendTelemetry from 'logic/telemetry/useSendTelemetry';
 import TimeRangeInputSettingsContext from 'views/components/contexts/TimeRangeInputSettingsContext';
+import type { TimeRangeType } from 'views/components/searchbar/time-range-picker/TimeRangeDropdown';
+import TimeRangeDropdown from 'views/components/searchbar/time-range-picker/TimeRangeDropdown';
 
-import TimeRangeDropdownButton from './TimeRangeDropdownButton';
-import type { TimeRangeType } from './date-time-picker/TimeRangeDropdown';
-import TimeRangeDropdown from './date-time-picker/TimeRangeDropdown';
+import TimeRangeDropdownButton from './time-range-preset/TimeRangePresetSelect';
 import TimeRangeDisplay from './TimeRangeDisplay';
 
 type Props = {
