+++ graylog2-web-interface/src/views/components/searchbar/date-time-picker/TabAbsoluteTimeRange.tsx
@@ -147,8 +147,11 @@ const TabAbsoluteTimeRange = ({ disabled, limitDuration }: Props) => {
           </AccordionItem>
         </StyledAccordion>
       </AbsoluteWrapper>
-      {showAbsolutePresetsButton
-  && (<TabPresetDropdown disabled={disabled} onSetPreset={onSetPreset} availableOptions={absoluteOptions} />)}
+      {showAbsolutePresetsButton && (
+        <TabPresetDropdown disabled={disabled}
+                           onSetPreset={onSetPreset}
+                           availableOptions={absoluteOptions} />
+      )}
     </>
   );
 };
