+++ graylog2-web-interface/src/views/components/searchbar/time-range-preset/TimeRangePresetSelect.tsx
@@ -23,8 +23,9 @@ import { ButtonGroup } from 'components/bootstrap';
 import { normalizeIfAllMessagesRange } from 'views/logic/queries/NormalizeTimeRange';
 import useSearchConfiguration from 'hooks/useSearchConfiguration';
 
-import RangePresetDropdown from './RangePresetDropdown';
-import TimeRangeButton from './TimeRangeButton';
+import RangePresetDropdown from './TimeRangePresetDropdown';
+
+import TimeRangeButton from '../TimeRangeButton';
 
 type Props = {
   disabled?: boolean,
@@ -44,7 +45,7 @@ const StyledButtonGroup = styled(ButtonGroup)`
   align-items: start;
 `;
 
-const TimeRangeDropdownButton = ({
+const TimeRangePresetSelect = ({
   disabled,
   hasErrorOnMount,
   onPresetSelectOpen,
@@ -80,24 +81,23 @@ const TimeRangeDropdownButton = ({
       <TimeRangeButton hasError={hasErrorOnMount}
                        disabled={disabled}
                        onClick={_onClick} />
-      {showPresetDropdown
-        && (
-          <StyledRangePresetDropdown disabled={disabled}
-                                     displayTitle={false}
-                                     onChange={selectRelativeTimeRangePreset}
-                                     onToggle={_onPresetSelectToggle}
-                                     header="Select time range"
-                                     bsSize={null}
-                                     availableOptions={availableOptions} />
-        )}
+      {showPresetDropdown && (
+        <StyledRangePresetDropdown disabled={disabled}
+                                   displayTitle={false}
+                                   onChange={selectRelativeTimeRangePreset}
+                                   onToggle={_onPresetSelectToggle}
+                                   header="Select time range"
+                                   bsSize={null}
+                                   availableOptions={availableOptions} />
+      )}
     </StyledButtonGroup>
   );
 };
 
-TimeRangeDropdownButton.defaultProps = {
+TimeRangePresetSelect.defaultProps = {
   hasErrorOnMount: false,
   disabled: false,
   showPresetDropdown: true,
 };
 
-export default TimeRangeDropdownButton;
+export default TimeRangePresetSelect;
